12.13. Фото на память. Перенос переменных.
