+++ js/filters.js
@@ -1,53 +1,5 @@
-const EFFECTS = [
-  {
-    name : 'none',
-    style : 'none',
-    min : 0 ,
-    max : 100,
-    step : 1,
-    unit : '',
-  },
-  {
-    name : 'chrome',
-    style : 'grayscale',
-    min : 0 ,
-    max : 1,
-    step : 0.1,
-    unit : '',
-  },
-  {
-    name : 'sepia',
-    style : 'sepia',
-    min : 0,
-    max : 1,
-    step : 0.1,
-    unit : '',
-  },
-  {
-    name : 'marvin',
-    style : 'invert',
-    min : 0,
-    max : 100,
-    step : 1,
-    unit : '%',
-  },
-  {
-    name : 'phobos',
-    style : 'blur',
-    min : 0,
-    max : 3,
-    step : 0.1,
-    unit : 'px',
-  },
-  {
-    name : 'heat',
-    style : 'brightness',
-    min : 1,
-    max : 3,
-    step : 0.1,
-    unit : '',
-  }
-];
+import {EFFECTS} from './data.js';
+
 const DEFAULT_EFFECT = EFFECTS[0];
 let currentEffect = DEFAULT_EFFECT;
 
@@ -59,7 +11,6 @@ const imagePreview = document.querySelector('.img-upload__preview img');
 
 const showSliderContainer = () => sliderContainer.classList.remove('hidden');
 const hideSliderContainer = () => sliderContainer.classList.add('hidden');
-
 
 const createSlider = () => {
   noUiSlider.create(sliderElement, {
