+++ js/form-validation.js
@@ -1,20 +1,8 @@
-const HASHTAG_MAX_COUNT = 5;
-const HASHTAG_MAX_LENGTH = 20;
-const COMMENT_MAX_LENGTH = 140;
-
-const HASHTAG_VALID = /^#[a-zа-яё0-9]{1,19}$/i;
-
-const ErrorMessages = {
-  INVALID_COUNT: `Максимум ${HASHTAG_MAX_COUNT} хэштегов`,
-  INVALID_HASHTAG: `Хештег должен начинаться с #, состоять из букв, чисел, не превышать ${HASHTAG_MAX_LENGTH} символов.`,
-  NOT_UNIQUE: 'Хэштеги должны быть уникальными',
-  INVALID_LENGTH_COMMENT: `Максимум ${COMMENT_MAX_LENGTH} символов`,
-};
+import {HASHTAG_MAX_COUNT, COMMENT_MAX_LENGTH, HASHTAG_VALID, ErrorMessages} from './data.js';
 
 const uploadForm = document.querySelector('.img-upload__form');
 const textHashtags = document.querySelector('.text__hashtags');
 const textDescription = document.querySelector('.text__description');
-
 
 const pristine = new Pristine(uploadForm ,{
   classTo: 'img-upload__field-wrapper',
