+++ js/gallery.js
@@ -3,20 +3,24 @@ import {renderPictures} from './thumbnails.js';
 
 const container = document.querySelector('.pictures');
 
-const renderGallery = (pictures) => {
-  container.addEventListener ('click', (evt) => {
-    const thumbnail = evt.target.closest ('[data-thumbnail-id]');
-    if (!thumbnail) {
-      return;
-    }
-    evt.preventDefault();
-    const picture = pictures.find (
-      (item) => item.id === +thumbnail.dataset.thumbnailId
-    );
-    showBigPicture(picture);
-  });
+let pictures = [];
 
+const onContainerClick = (evt) => {
+  const thumbnail = evt.target.closest ('[data-thumbnail-id]');
+  if (!thumbnail) {
+    return;
+  }
+  evt.preventDefault();
+  const picture = pictures.find (
+    (item) => item.id === +thumbnail.dataset.thumbnailId
+  );
+  showBigPicture(picture);
+};
+
+const renderGallery = (currentPictures) => {
+  pictures = currentPictures;
   renderPictures (pictures, container);
+  container.addEventListener ('click', onContainerClick);
 };
 
 export {renderGallery};
