+++ js/img-upload-file.js
@@ -1,0 +1,23 @@
+import {FILE_TYPES} from './data.js';
+
+const fileChooser = document.querySelector('.img-upload__input[type=file]');
+const preview = document.querySelector('.img-upload__preview img');
+const previewEffects = document.querySelectorAll('.effects__preview');
+
+const addFileChooserListener = () => {
+  fileChooser.addEventListener('change', () => {
+    const file = fileChooser.files[0];
+    const fileName = file.name.toLowerCase();
+
+    const matches = FILE_TYPES.some((it) => fileName.endsWith(it));
+
+    if (matches) {
+      preview.src = URL.createObjectURL(file);
+      previewEffects.forEach((value) => {
+        value.style.backgroundImage = `url(${preview.src})`;
+      });
+    }
+  });
+};
+
+export {addFileChooserListener};
