+++ js/main.js
@@ -3,19 +3,21 @@ import {renderModalForm} from './form.js';
 import {setDefaultSlider} from './filters.js';
 import {getData} from './api.js';
 import {setUserFormSubmit} from './form.js';
-import {showAlert} from './utils.js';
+import {debounce, showAlert} from './utils.js';
 import {showFilters} from './img-filters.js';
+import {addFileChooserListener} from './img-upload-file.js';
 
 getData().then((picturesArray)=>{
-  renderGallery(picturesArray);
-  showFilters(picturesArray);
-  renderModalForm(picturesArray);
-})
-  .catch(
-    (err) => {
+  const debouncedRenderGallery = debounce (renderGallery);
+   renderGallery(picturesArray);
+   showFilters(picturesArray, debouncedRenderGallery);
+   renderModalForm(picturesArray)
+ })
+   .catch((err) => {
       showAlert(err.message);
     }
-  );
+   );
 
 setDefaultSlider();
 setUserFormSubmit();
+addFileChooserListener();
